Split Docker-independent parts of dockerImage into docker/image.go
This does not change the code at all, only moving things around now.
